USER-SIDE Profile Beta 1.30
Added UserReviewsController.java in controller files.

Added and Updated lines in DatabaseHandler for getReviews for the User

Reviews now working as intended, when clicking the button reviews in the profile section it'll now display the FXML file of review then showing the user's reviews with poster on his watched movies

Watchlist is working properly as intended (Validated).

Errors:

The size of the poster is still not accurate to the figma basis file

The username and bio is wonky, need to be updated in the next patch