USER-SIDE Profile-Ratings-Review BETA Final
Tested with a few contents on my database, I highly suggest that we do a thorough validation/testing of this USER-SIDE. But I'm feeling confident enough with my work.

Renamed some of the variables, methods, queries to match the esthetic? or the format. Haven't clean all the comments for future studying/revisiting this code.